Refactor test fixtures and parameters,
remove redundant code, update environment variables.
